+++ frictionless/plugins/spss/__init__.py
@@ -1,0 +1,3 @@
+from .plugin import SpssPlugin
+from .dialect import SpssDialect
+from .parser import SpssParser
